added more inventory functionality
long press on an entry to edit attributes
add item activity for new entries
clear table activity
price data stored as INT

+++ app/src/main/java/org/csuf/cpsc411/cpsc411project/Item.kt
@@ -6,17 +6,17 @@ class Item {
     var itemId : Int = 0
     var itemName : String = "Default_Name"
     var itemQty : Int = 0
-    //TODO: Change itemPrice to be an Integer stored in the SQLite database for accuracy, then use BigDecimal in backend
-    var itemPrice : Float = 0.00F
+    // itemPrice is stored as an Int: smallest unit of currency (cost in cents)
+    var itemPrice : Int = 0
 
-    constructor(itemId: Int, itemName: String, itemQty: Int, itemPrice: Float ){
+    constructor(itemId: Int, itemName: String, itemQty: Int, itemPrice: Int ){
         this.itemId = itemId
         this.itemName = itemName
         this.itemQty = itemQty
         this.itemPrice = itemPrice
     }
 
-    constructor(itemName: String, itemQty: Int, itemPrice: Float){
+    constructor(itemName: String, itemQty: Int, itemPrice: Int){
         this.itemName = itemName
         this.itemQty = itemQty
         this.itemPrice = itemPrice
